rename api to graphql_api
