+++ platformics/codegen/templates/graphql_api/helpers/class_name.py.j2
@@ -2,7 +2,7 @@
 Define GraphQL types and helper functions for supporting GROUPBY queries.
 
 Auto-gereanted by running 'make codegen'. Do not edit.
-Make changes to the template codegen/templates/api/groupby_helpers.py.j2 instead.
+Make changes to the template codegen/templates/graphql_api/groupby_helpers.py.j2 instead.
 """
 {% set related_fields = cls.related_fields | unique(attribute='related_class.name') | list %}
 {% set ignored_fields = ["File", "Entity", cls.name] %}
@@ -23,7 +23,7 @@ from support.enums import
 
 {%- for field in cls.related_fields %}
     {%- if field.related_class.name not in ignored_fields and not field.multivalued and not field.is_virtual_relationship %}
-from api.helpers.{{ field.related_class.snake_name }} import {{ field.related_class.name }}GroupByOptions, build_{{ field.related_class.snake_name }}_groupby_output
+from graphql_api.helpers.{{ field.related_class.snake_name }} import {{ field.related_class.name }}GroupByOptions, build_{{ field.related_class.snake_name }}_groupby_output
     {%- endif %}
 {%- endfor %}
