+++ platformics/codegen/templates/graphql_api/mutations.py.j2
@@ -2,17 +2,17 @@
 GraphQL mutations for files and entities
 
 Auto-generated by running 'make codegen'. Do not edit.
-Make changes to the template codegen/templates/api/mutations.py.j2 instead.
+Make changes to the template codegen/templates/graphql_api/mutations.py.j2 instead.
 """
 
 import strawberry
 from typing import Sequence
 {%- if render_files %}
-from platformics.api.files import File, create_file, upload_file, upload_temporary_file, mark_upload_complete, concatenate_files, SignedURL, MultipartUploadResponse
+from platformics.graphql_api.files import File, create_file, upload_file, upload_temporary_file, mark_upload_complete, concatenate_files, SignedURL, MultipartUploadResponse
 {%- endif %}
 
 {%- for class in classes %}
-from api.types.{{ class.snake_name }} import {{ class.name }}, {%- if class.create_fields %}create_{{ class.snake_name }}, {%- endif %}{%- if class.mutable_fields %}update_{{ class.snake_name }}, {%- endif %}delete_{{ class.snake_name }}
+from graphql_api.types.{{ class.snake_name }} import {{ class.name }}, {%- if class.create_fields %}create_{{ class.snake_name }}, {%- endif %}{%- if class.mutable_fields %}update_{{ class.snake_name }}, {%- endif %}delete_{{ class.snake_name }}
 {%- endfor %}
 
 @strawberry.type
