+++ platformics/codegen/templates/graphql_api/queries.py.j2
@@ -2,17 +2,17 @@
 Supported GraphQL queries for files and entities
 
 Auto-generated by running 'make codegen'. Do not edit.
-Make changes to the template codegen/templates/api/queries.py.j2 instead.
+Make changes to the template codegen/templates/graphql_api/queries.py.j2 instead.
 """
 
 import strawberry
-from platformics.api import relay
+from platformics.graphql_api import relay
 from typing import Sequence, List
 {%- if render_files %}
-from platformics.api.files import File, resolve_files
+from platformics.graphql_api.files import File, resolve_files
 {%- endif %}
 {%- for class in classes %}
-from api.types.{{ class.snake_name }} import {{ class.name }}, resolve_{{ class.plural_snake_name }}, {{ class.name }}Aggregate, resolve_{{ class.plural_snake_name }}_aggregate
+from graphql_api.types.{{ class.snake_name }} import {{ class.name }}, resolve_{{ class.plural_snake_name }}, {{ class.name }}Aggregate, resolve_{{ class.plural_snake_name }}_aggregate
 {%- endfor %}
