+++ platformics/graphql_api/core/deps.py
@@ -10,7 +10,7 @@ from mypy_boto3_sts.client import STSClient
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.requests import Request
 
-from platformics.api.core.error_handler import PlatformicsError
+from platformics.graphql_api.core.error_handler import PlatformicsError
 from platformics.database.connect import AsyncDB, init_async_db
 from platformics.security.token_auth import get_token_claims
 from platformics.settings import APISettings
