+++ platformics/graphql_api/core/query_builder.py
@@ -17,8 +17,8 @@ from sqlalchemy.sql import Select
 from typing_extensions import TypedDict
 
 import platformics.database.models as db
-from platformics.api.core.errors import PlatformicsError
-from platformics.api.core.query_input_types import aggregator_map, operator_map, orderBy
+from platformics.graphql_api.core.errors import PlatformicsError
+from platformics.graphql_api.core.query_input_types import aggregator_map, operator_map, orderBy
 from platformics.database.models.base import Base
 from platformics.security.authorization import CerbosAction, get_resource_query
