+++ test_app/etc/supervisord.conf
@@ -33,7 +33,7 @@ startretries = 9999
 
 [program:graphql_api]
 directory=/app
-command=gunicorn --reload -k platformics.api.uvicorn_worker.ReloadingWorker -c /app/etc/gunicorn_conf.py main:app
+command=gunicorn --reload -k platformics.graphql_api.uvicorn_worker.ReloadingWorker -c /app/etc/gunicorn_conf.py main:app
 environment=PYTHONUNBUFFERED=1,PYTHONPATH=.
 stderr_logfile = /dev/stderr
 stderr_logfile_maxbytes=0
